Fix NumericApplicationTransactionCounterTrack1 implicit operator for underlying _Value.

+++ Source/Emvco/Emv/Play.Emv.Ber/DataElements/Card/Magstripe/NumericApplicationTransactionCounterTrack1.cs
@@ -54,7 +54,7 @@ public record NumericApplicationTransactionCounterTrack1 : DataElement<byte>
 
     #region Operator Overrides
 
-    public static implicit operator short(NumericApplicationTransactionCounterTrack1 value) => value._Value;
+    public static implicit operator byte(NumericApplicationTransactionCounterTrack1 value) => value._Value;
 
     #endregion
